Delete separeted feature and Rename
